+++ src/Extensions/RenderedFragmentQueryExtensions.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -31,16 +30,18 @@ namespace Bunit
         }
 
         /// <summary>
-        /// Returns a list of elements from the rendered fragment or component under test, 
+        /// Returns a refreshable collection of <see cref="IElement"/>s from the rendered fragment or component under test, 
         /// using the provided <paramref name="cssSelector"/>, in a depth-first pre-order traversal 
         /// of the rendered nodes.
         /// </summary>
         /// <param name="renderedFragment">The rendered fragment to search.</param>
         /// <param name="cssSelector">The group of selectors to use.</param>
-        public static IHtmlCollection<IElement> FindAll(this IRenderedFragment renderedFragment, string cssSelector)
+        /// <param name="enableAutoRefresh">If true, the returned <see cref="IRefreshableElementCollection{IElement}"/> will automatically refresh its <see cref="IElement"/>s whenever the <paramref name="renderedFragment"/> changes.</param>
+        /// <returns>An <see cref="IRefreshableElementCollection{IElement}"/>, that can be refreshed to execute the search again.</returns>
+        public static IRefreshableElementCollection<IElement> FindAll(this IRenderedFragment renderedFragment, string cssSelector, bool enableAutoRefresh = false)
         {
             if (renderedFragment is null) throw new ArgumentNullException(nameof(renderedFragment));
-            return renderedFragment.Nodes.QuerySelectorAll(cssSelector);
+            return new RefreshableElementCollection(renderedFragment, cssSelector) { EnableAutoRefresh = enableAutoRefresh };
         }
     }
 }